feat: Clean up ShareExtension
